refactor(discord-bot): organize command handling
