+++ bots/discord-bot-http/src/utils/generalComponents.ts
@@ -1,22 +1,18 @@
 import {
   DISCORD_APPLICATION_COMMANDS,
+  DISCORD_BUTTON_TEST_ACTIONS,
+  DISCORD_BUTTON_TEST_CUSTOM_ID_PREFIX,
   DISCORD_BUTTON_STYLES,
   DISCORD_COMPONENT_TYPES,
   DISCORD_EMBED_COLORS,
   DISCORD_MESSAGE_FLAGS,
 } from "../constants/discord.js";
-import type { DiscordInteractionResponseData } from "../types/discord.js";
-
-const BUTTON_TEST_CUSTOM_ID_PREFIX = "button-test";
-
-const BUTTON_TEST_ACTIONS = {
-  PRIMARY: "primary",
-  SECONDARY: "secondary",
-  SUCCESS: "success",
-  DANGER: "danger",
-} as const;
-
-type ButtonTestAction = (typeof BUTTON_TEST_ACTIONS)[keyof typeof BUTTON_TEST_ACTIONS];
+import type {
+  ButtonTestAction,
+  DiscordInfoResponseInput,
+  DiscordInteractionResponseData,
+  DiscordProfileResponseInput,
+} from "../types/discord.js";
 
 const withComponentsV2Flags = (
   data: Omit<DiscordInteractionResponseData, "flags">,
@@ -49,32 +45,33 @@ const createContainer = (
   components,
 });
 
-const createButtonTestCustomId = (action: ButtonTestAction): string => `${BUTTON_TEST_CUSTOM_ID_PREFIX}:${action}`;
+const createButtonTestCustomId = (action: ButtonTestAction): string =>
+  `${DISCORD_BUTTON_TEST_CUSTOM_ID_PREFIX}:${action}`;
 
 const createButtonTestActionRow = () => ({
   type: DISCORD_COMPONENT_TYPES.ACTION_ROW,
   components: [
     {
       type: DISCORD_COMPONENT_TYPES.BUTTON,
-      custom_id: createButtonTestCustomId(BUTTON_TEST_ACTIONS.PRIMARY),
+      custom_id: createButtonTestCustomId(DISCORD_BUTTON_TEST_ACTIONS.PRIMARY),
       label: "Refresh",
       style: DISCORD_BUTTON_STYLES.PRIMARY,
     },
     {
       type: DISCORD_COMPONENT_TYPES.BUTTON,
-      custom_id: createButtonTestCustomId(BUTTON_TEST_ACTIONS.SECONDARY),
+      custom_id: createButtonTestCustomId(DISCORD_BUTTON_TEST_ACTIONS.SECONDARY),
       label: "Echo",
       style: DISCORD_BUTTON_STYLES.SECONDARY,
     },
     {
       type: DISCORD_COMPONENT_TYPES.BUTTON,
-      custom_id: createButtonTestCustomId(BUTTON_TEST_ACTIONS.SUCCESS),
+      custom_id: createButtonTestCustomId(DISCORD_BUTTON_TEST_ACTIONS.SUCCESS),
       label: "Success",
       style: DISCORD_BUTTON_STYLES.SUCCESS,
     },
     {
       type: DISCORD_COMPONENT_TYPES.BUTTON,
-      custom_id: createButtonTestCustomId(BUTTON_TEST_ACTIONS.DANGER),
+      custom_id: createButtonTestCustomId(DISCORD_BUTTON_TEST_ACTIONS.DANGER),
       label: "Danger",
       style: DISCORD_BUTTON_STYLES.DANGER,
     },
@@ -115,14 +112,7 @@ export const createInfoResponseData = ({
   method,
   platform,
   runtime,
-}: {
-  branch: string;
-  ephemeral: boolean;
-  functionUrl: string;
-  method: string;
-  platform: string;
-  runtime: string;
-}): DiscordInteractionResponseData =>
+}: DiscordInfoResponseInput): DiscordInteractionResponseData =>
   withComponentsV2Flags(
     {
       components: [
@@ -149,12 +139,7 @@ export const createProfileResponseData = ({
   name,
   totalRuns,
   usageLines,
-}: {
-  ephemeral: boolean;
-  name: string | undefined;
-  totalRuns: number;
-  usageLines: string[];
-}): DiscordInteractionResponseData =>
+}: DiscordProfileResponseInput): DiscordInteractionResponseData =>
   withComponentsV2Flags(
     {
       components: [
@@ -199,11 +184,11 @@ export const createButtonTestResponseData = (
 export const parseButtonTestCustomId = (customId: string): ButtonTestAction | undefined => {
   const [prefix, action] = customId.split(":");
 
-  if (prefix !== BUTTON_TEST_CUSTOM_ID_PREFIX) {
+  if (prefix !== DISCORD_BUTTON_TEST_CUSTOM_ID_PREFIX) {
     return undefined;
   }
 
-  return Object.values(BUTTON_TEST_ACTIONS).includes(action as ButtonTestAction)
+  return Object.values(DISCORD_BUTTON_TEST_ACTIONS).includes(action as ButtonTestAction)
     ? (action as ButtonTestAction)
     : undefined;
 };
@@ -212,13 +197,13 @@ export const createButtonTestClickResponseData = (action: ButtonTestAction): Dis
   const clickedAt = Math.floor(Date.now() / 1000);
 
   switch (action) {
-    case BUTTON_TEST_ACTIONS.PRIMARY:
+    case DISCORD_BUTTON_TEST_ACTIONS.PRIMARY:
       return createButtonTestResponseData(false, `Refreshed <t:${clickedAt}:R>.`);
-    case BUTTON_TEST_ACTIONS.SECONDARY:
+    case DISCORD_BUTTON_TEST_ACTIONS.SECONDARY:
       return createButtonTestResponseData(false, "Echo: the secondary button was clicked.");
-    case BUTTON_TEST_ACTIONS.SUCCESS:
+    case DISCORD_BUTTON_TEST_ACTIONS.SUCCESS:
       return createButtonTestResponseData(false, "Success: the green button completed its action.");
-    case BUTTON_TEST_ACTIONS.DANGER:
+    case DISCORD_BUTTON_TEST_ACTIONS.DANGER:
       return createButtonTestResponseData(false, "Danger: the red button was clicked. Nothing destructive happened.");
   }
 };
